ru.progwards.java1.lessons.basics

+++ src/ru/progwards/java1/lessons/compare_if_cycles/CyclesGoldenFibo.java
@@ -2,7 +2,7 @@ package ru.progwards.java1.lessons.compare_if_cycles;
 
 public class CyclesGoldenFibo {
     public static boolean containsDigit(int number, int digit){
-        int num=number/10;
+        float num=number/10;
         if(number%10==digit||num%10==digit) return true;
         else return false;
     }
